+++ examples/graphql-annotations/index.php
@@ -3,14 +3,14 @@
 namespace App;
 
 use Siler\GraphQL\DateTimeScalar;
-use function Siler\GraphQL\annotated;
-use function Siler\GraphQL\debug;
-use function Siler\GraphQL\init;
+use function Siler\GraphQL\{debug, init};
+use function Siler\GraphQL\Annotation\annotated;
 
 require_once __DIR__ . '/vendor/autoload.php';
 require_once __DIR__ . '/../../vendor/autoload.php';
 
 $schema = annotated([
+    TodoStatus::class,
     Todo::class,
     TupleInput::class,
     Query::class,
